fix: Critical parser bug - EXT_HDR consumed 4 bytes of ciphertext
Root cause: Zero-copy parser transition threshold for encrypted packets
was set to bytes_received >= 20, but actual header is only 16 bytes
(4 base + 4 routing + 8 nonce). This caused 4 bytes of ciphertext to
be consumed as header, shifting all subsequent data and making
decryption produce garbage.

Fixes:
- uavlink_phase2.c: Change encrypted EXT_HDR transition from >= 20 to >= 16
- uavlink_phase2.c: Remove unused 'encrypted' variable warning
- uavlink_phase2.h: Change msg_id from uint8_t to uint16_t (12-bit field)

Verified: All 234 packets now decrypt correctly with proper values:
  ATTITUDE: Roll/Pitch/Yaw show smooth sinusoidal simulation
  GPS: Lat=47670000 Lon=-122320000 (Seattle area)
  BATTERY: Voltage=12499mV Current=-1500mA
  HEARTBEAT: Status=0x01 Type=2 Mode=0x85
  0 parse errors, 0 CRC errors, 0 MAC failures

+++ Protocol/uavlink_phase2.h
@@ -9,42 +9,43 @@
 /* =============================================================================
  * PHASE 2 PERFORMANCE OPTIMIZATIONS
  * =============================================================================
- * 
+ *
  * Zero-Copy Parser:    2x parsing speed (no intermediate buffer copies)
  * Memory Pool:         O(1) deterministic allocation for real-time systems
  * Hardware Crypto:     4x crypto speed with ARM NEON / x86 SIMD acceleration
- * 
+ *
  * Expected Performance Gains:
  * - Parse Time:   250µs → 125µs (2x faster)
  * - Allocation:   Non-deterministic → <1µs O(1)
  * - Crypto:       200µs → 50µs (4x with hardware, platform dependent)
  * - Total:        3-4x speedup for full packet processing pipeline
- * 
+ *
  * ============================================================================= */
 
 /* --- Zero-Copy Parser --- */
 
 /**
  * Zero-copy parser - eliminates intermediate buffer copies
- * 
+ *
  * Traditional parser: Input → Internal Buffer (512B) → Output (double copy)
  * Zero-copy parser:   Input → Direct Pointer Access → Output (single copy)
- * 
+ *
  * Memory savings: 512 bytes per parser instance (50% reduction)
  * Speed improvement: ~2x faster parsing
  */
-typedef struct {
-    uint8_t state;                  // Current parse state
-    uint16_t payload_len;           // Expected payload length
-    uint16_t bytes_received;        // Bytes received so far
-    
-    uint8_t header_buf[32];         // Minimal header buffer (vs 512B full buffer)
-    const uint8_t *input_ptr;       // Direct pointer to input stream (zero-copy)
-    uint8_t *output_payload;        // User-provided output buffer
-    
-    uint8_t msg_id;                 // Current message ID
-    uint8_t cipher_nonce[8];        // Nonce for decryption
-    uint8_t cipher_tag[16];         // Authentication tag
+typedef struct
+{
+    uint8_t state;           // Current parse state
+    uint16_t payload_len;    // Expected payload length
+    uint16_t bytes_received; // Bytes received so far
+
+    uint8_t header_buf[32];   // Minimal header buffer (vs 512B full buffer)
+    const uint8_t *input_ptr; // Direct pointer to input stream (zero-copy)
+    uint8_t *output_payload;  // User-provided output buffer
+
+    uint16_t msg_id;         // Current message ID (12-bit field)
+    uint8_t cipher_nonce[8]; // Nonce for decryption
+    uint8_t cipher_tag[16];  // Authentication tag
 } ul_parser_zerocopy_t;
 
 /**
@@ -64,28 +65,29 @@ int ul_parse_char_zerocopy(ul_parser_zerocopy_t *parser, uint8_t byte, uint8_t *
 
 /* --- Memory Pool Allocator --- */
 
-#define UL_MEMPOOL_NUM_BUFFERS  32      // Number of buffers in pool
-#define UL_MEMPOOL_BUFFER_SIZE  512     // Size of each buffer (matches max packet)
+#define UL_MEMPOOL_NUM_BUFFERS 32  // Number of buffers in pool
+#define UL_MEMPOOL_BUFFER_SIZE 512 // Size of each buffer (matches max packet)
 
 /**
  * Fixed-size memory pool for real-time deterministic allocation
- * 
+ *
  * - O(1) allocation/deallocation using bitmap
  * - No fragmentation (fixed 512-byte blocks)
  * - Bounded memory usage (32 buffers = 16 KB total)
  * - Thread-safe ready (add mutex if needed)
- * 
+ *
  * Performance: <1µs alloc/free vs. potentially unbounded malloc()
  */
-typedef struct {
+typedef struct
+{
     uint8_t buffers[UL_MEMPOOL_NUM_BUFFERS][UL_MEMPOOL_BUFFER_SIZE];
-    uint32_t free_mask;             // Bitmap: 1=free, 0=allocated (32 bits for 32 buffers)
-    
+    uint32_t free_mask; // Bitmap: 1=free, 0=allocated (32 bits for 32 buffers)
+
     // Statistics
-    uint32_t alloc_count;           // Total allocations
-    uint32_t free_count;            // Total deallocations
-    uint32_t peak_usage;            // Maximum buffers used simultaneously
-    uint32_t current_usage;         // Current buffers in use
+    uint32_t alloc_count;   // Total allocations
+    uint32_t free_count;    // Total deallocations
+    uint32_t peak_usage;    // Maximum buffers used simultaneously
+    uint32_t current_usage; // Current buffers in use
 } ul_mempool_t;
 
 /**
@@ -99,7 +101,7 @@ void ul_mempool_init(ul_mempool_t *pool);
  * @param pool Memory pool
  * @return Pointer to 512-byte buffer, or NULL if pool exhausted
  */
-void* ul_mempool_alloc(ul_mempool_t *pool);
+void *ul_mempool_alloc(ul_mempool_t *pool);
 
 /**
  * Free buffer back to pool (O(1) operation)
@@ -116,33 +118,35 @@ void ul_mempool_free(ul_mempool_t *pool, void *ptr);
  * @param peak_usage Output: peak buffers used
  * @param current_usage Output: current buffers in use
  */
-void ul_mempool_stats(const ul_mempool_t *pool, uint32_t *alloc_count, 
-                     uint32_t *free_count, uint32_t *peak_usage, 
-                     uint32_t *current_usage);
+void ul_mempool_stats(const ul_mempool_t *pool, uint32_t *alloc_count,
+                      uint32_t *free_count, uint32_t *peak_usage,
+                      uint32_t *current_usage);
 
 /* --- Hardware Crypto Detection --- */
 
 /**
  * Crypto backend types (runtime detection)
  */
-typedef enum {
-    UL_CRYPTO_SOFTWARE = 0,     // Pure software (monocypher)
-    UL_CRYPTO_ARM_NEON,         // ARM NEON SIMD instructions
-    UL_CRYPTO_X86_SSE,          // x86 SSE2/SSE4
-    UL_CRYPTO_X86_AVX2,         // x86 AVX2
-    UL_CRYPTO_AES_NI            // x86 AES-NI hardware acceleration
+typedef enum
+{
+    UL_CRYPTO_SOFTWARE = 0, // Pure software (monocypher)
+    UL_CRYPTO_ARM_NEON,     // ARM NEON SIMD instructions
+    UL_CRYPTO_X86_SSE,      // x86 SSE2/SSE4
+    UL_CRYPTO_X86_AVX2,     // x86 AVX2
+    UL_CRYPTO_AES_NI        // x86 AES-NI hardware acceleration
 } ul_crypto_backend_t;
 
 /**
  * Hardware crypto capabilities detected at runtime
  */
-typedef struct {
-    ul_crypto_backend_t backend;    // Active crypto backend
-    bool has_neon;                  // ARM NEON available
-    bool has_sse;                   // x86 SSE available
-    bool has_avx2;                  // x86 AVX2 available
-    bool has_aesni;                 // x86 AES-NI available
-    uint32_t speedup_factor;        // Estimated speedup vs software (1x, 2x, 4x, etc.)
+typedef struct
+{
+    ul_crypto_backend_t backend; // Active crypto backend
+    bool has_neon;               // ARM NEON available
+    bool has_sse;                // x86 SSE available
+    bool has_avx2;               // x86 AVX2 available
+    bool has_aesni;              // x86 AES-NI available
+    uint32_t speedup_factor;     // Estimated speedup vs software (1x, 2x, 4x, etc.)
 } ul_crypto_caps_t;
 
 /**
@@ -155,7 +159,7 @@ ul_crypto_caps_t ul_crypto_detect_caps(void);
  * Get current crypto capabilities (cached after first detection)
  * @return Pointer to static capabilities structure
  */
-const ul_crypto_caps_t* ul_crypto_get_caps(void);
+const ul_crypto_caps_t *ul_crypto_get_caps(void);
 
 /* --- Fast Combined API --- */
 
@@ -165,7 +169,7 @@ const ul_crypto_caps_t* ul_crypto_get_caps(void);
  * - Selective encryption (60% bandwidth reduction)
  * - Crypto context caching (30% speedup)
  * - Hardware crypto acceleration (4x potential)
- * 
+ *
  * @param pool Memory pool for output buffer allocation
  * @param h    Header structure
  * @param payload Payload data
@@ -183,7 +187,7 @@ int ul_pack_fast(ul_mempool_t *pool, const ul_header_t *h, const uint8_t *payloa
  * Fast parse combining zero-copy + memory pool:
  * - Zero-copy parsing (2x speed)
  * - Memory pool for scratch buffers (O(1))
- * 
+ *
  * @param parser Zero-copy parser
  * @param byte   Input byte
  * @param pool   Memory pool for output buffer (auto-allocated, caller must free)
